+++ tests/test_selectors.py
@@ -72,7 +72,7 @@ def test_with_memory_priority_handling():
     assert_banner()
     root = py_trees.composites.Selector(name="Selector w/ Memory", memory=True)
     child_1 = py_trees.behaviours.Failure("Failure")
-    child_2 = py_trees.behaviours.StatusSequence(
+    child_2 = py_trees.behaviours.StatusQueue(
         name="Run-Fail",
         sequence=[py_trees.common.Status.RUNNING,
                   py_trees.common.Status.FAILURE],